remove flag
