+++ pkg/apis/pingcap/v1alpha1/validation/validation.go
@@ -209,8 +209,7 @@ func validateStartScriptFeatureFlags(featureFlags []v1alpha1.StartScriptV2Featur
 	for i, ff := range featureFlags {
 		idxPath := fldPath.Index(i)
 		if ff != v1alpha1.StartScriptV2FeatureFlagWaitForDnsNameIpMatch &&
-			ff != v1alpha1.StartScriptV2FeatureFlagPreferPDAddressesOverDiscovery &&
-			ff != v1alpha1.StartScriptV2FeatureFlagPDMSWithName {
+			ff != v1alpha1.StartScriptV2FeatureFlagPreferPDAddressesOverDiscovery {
 			allErrs = append(allErrs, field.Invalid(idxPath, ff, "Invalid start script feature flag"))
 		}
 	}
